Utility functions used for various plotting routines within atlantistools.

Utility functions used for various plotting routines within atlantistools.

Usage

custom_grid(plot, grid_x, grid_y)

Arguments

plot

ggplot2 plot.

grid_x

character vector used in facet_grid in horizontal direction.

grid_y

character vector used in facet_grid in vertical direction.

Value

ggplot2 plot.