Get boundary boxes from Atlantis box information.

Use the output from load_box and obtain a vector specifying which boxes are along the boundary.

get_boundary(boxinfo)

Arguments

boxinfo

A list as returned from load_box.

Value

A vector specifying which boxes are on the boundary.

See also

Author

Kelli Faye Johnson

Examples

d <- system.file("extdata", "setas-model-new-trunk", package = "atlantistools")
boxes <- load_box(bgm = file.path(d, "VMPA_setas.bgm"))
get_boundary(boxinfo = boxes)
#> [1]  0  6  7  8  9 10