Extracts the names of the epibenthic biomasspools from the initial conditions file.

Use fgs data.frame as read in by load_fgs to get the biomass pool information.

Usage

load_bps(fgs, init)

Arguments

fgs

Character string giving the connection to the functional groups file. The filename usually contains Groups and does end in .csv.

init

Character string giving the connection of the initial conditions netcdf file. The filename usually contains init and ends in .nc.

Value

Character vector of epibenthic biomass pools.

See also

Examples

d <- system.file("extdata", "setas-model-new-trunk", package = "atlantistools")

fgs <- file.path(d, "SETasGroupsDem_NoCep.csv")
init <- file.path(d, "INIT_VMPA_Jan2015.nc")

load_bps(fgs, init)
#> [1] "Megazoobenthos"